Added Exception handles

+++ job/views.py
@@ -1,9 +1,10 @@
 from django.db.models import Avg, Count, Max, Min
 from django.shortcuts import get_object_or_404
 from rest_framework import status
-from rest_framework.decorators import api_view
-from rest_framework.response import Response
+from rest_framework.decorators import api_view, permission_classes
 from rest_framework.pagination import PageNumberPagination
+from rest_framework.permissions import IsAuthenticated
+from rest_framework.response import Response
 
 from .filters import JobsFilter
 from .models import Job
@@ -36,7 +37,9 @@ def getJob(request,pk):
     return Response(serializer.data, status=status.HTTP_200_OK)
 
 @api_view(['POST'])
+@permission_classes([IsAuthenticated])
 def newJob(request):
+    request.data['user']=request.user
     data = request.data
     job = Job.objects.create(**data)
 
@@ -44,8 +47,12 @@ def newJob(request):
     return Response(serializer.data, status=status.HTTP_200_OK)
 
 @api_view(['PUT'])
+@permission_classes([IsAuthenticated])
 def updateJob(request,pk):
     job = get_object_or_404(Job, id=pk)
+
+    if job.user != request.user:
+        return Response({'message': 'You can not update this job'}, status=status.HTTP_403_FORBIDDEN)
 
     job.title = request.data['title']
     job.description = request.data['description']
@@ -66,8 +73,12 @@ def updateJob(request,pk):
 
 
 @api_view(['DELETE'])
+@permission_classes([IsAuthenticated])
 def deleteJob(request, pk):
     job = get_object_or_404(Job, id=pk)
+    if job.user != request.user:
+        return Response({'message': 'You can not delete this job'}, status=status.HTTP_403_FORBIDDEN)
+
     job.delete()
 
     return Response({'message': 'Job is Deleted.'}, status=status.HTTP_200_OK)
